requirements.txt
APScheduler==3.6.1
asn1crypto==0.24.0
attrs==19.1.0
Automat==0.7.0
awsebcli==3.15.3
blessed==1.15.0
botocore==1.12.203
cached-property==1.5.1
cement==2.8.2
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
colorama==0.3.9
constantly==15.1.0
cryptography==2.7
cssselect==1.0.3
Django==2.2.4
docker==3.7.3
docker-compose==1.23.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
future==0.16.0
hyperlink==19.0.0
idna==2.7
incremental==17.5.0
jmespath==0.9.4
jsonschema==2.6.0
lxml==4.3.3
parsel==1.5.1
pathspec==0.5.9
pkg-resources==0.0.0
psycopg2==2.8.2
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pyOpenSSL==19.0.0
python-dateutil==2.8.0
python-docx==0.8.11
pytz==2019.1
PyYAML==3.13
queuelib==1.5.0
requests==2.20.1
Scrapy==1.6.0
semantic-version==2.5.0
service-identity==18.1.0
six==1.11.0
sqlparse==0.3.0
termcolor==1.1.0
texttable==0.9.1
Twisted==19.2.1
tzlocal==2.0.0
urllib3==1.24.3
w3lib==1.20.0
wcwidth==0.1.7
websocket-client==0.56.0
whitenoise==4.1.3
zope.interface==4.6.0
